Hmm
Looks like [the issue](https://github.com/smtech/canvas-ics-sync/issues/15) is/was that I was using the wrong field name for the sync setting. I'm pretty sure that this is all held together with bubble gum and good intentions — which is why I'm rewriting it from scratch. Meanwhile, this almost definitely breaks custom cron schedules (but who cares?), while fixing the "normal" schedules.